+++ tomcat-base/src/shared/main/java/eu/xenit/alfresco/tomcat/embedded/alfresco/config/AlfrescoConfigurationProvider.java
@@ -1,0 +1,13 @@
+package eu.xenit.alfresco.tomcat.embedded.alfresco.config;
+
+
+public interface AlfrescoConfigurationProvider {
+
+    default AlfrescoConfiguration getConfiguration() {
+        return getConfiguration(new AlfrescoConfiguration());
+    }
+
+    AlfrescoConfiguration getConfiguration(AlfrescoConfiguration baseConfiguration);
+
+
+}