+++ tomcat-base/src/shared/main/java/eu/xenit/alfresco/tomcat/embedded/share/config/ShareConfigurationProvider.java
@@ -1,0 +1,10 @@
+package eu.xenit.alfresco.tomcat.embedded.share.config;
+
+public interface ShareConfigurationProvider {
+
+    default ShareConfiguration getConfiguration() {
+        return getConfiguration(new ShareConfiguration());
+    }
+
+    ShareConfiguration getConfiguration(ShareConfiguration baseShareConfiguration);
+}